Battig - Montague Categorized Word Norms

This dataset comprises a ranked list of 5231 words listed in 56 taxonomic categories by people who were asked to list as many exemplars of a given category ("a precious stone", "a unit of time", "a fruit", "a color", etc.). Participants had 30s to generate as many responses to each category as possible, after which time the next category name was presented.

Included in this dataset are all words from the Battig and Montague (1969) norms listed with freq > 1.

Usage

data(Battig)

Format

A data frame with 5231 observations on the following 9 variables.

word

a character vector

catnum

category number, a factor

catname

category name, a factor

syl

number of syllables

letters

number of letters

freq

Frequency of response

frequency

Kucera-Francis word frequency

rank

rank of freq within the category

rfreq

rated frequency

Details

In our original dataset, words were truncated at 18 characters, so some are incomplete.

Source

Battig, W.F. & Montague, W.E. (1968). Category norms for verbal items in 56 categories: A replication and extension of the Connecticut norms using University of Maryland and Illinois students (Tech. Rep.) University of Colorado, Boulder, CO (1968)

Battig, W.F. & Montague, W.E. (1969). Category norms for verbal items in 56 categories: A replication and extension of the Connecticut norms. Journal of Experimental Psychology, 80 (1969), pp. 1-46

References

Joelson, J. M. & Hermann, D. J., Properties of categories in semantic | memory, Amer Journal of Psychology, 1978, 91, 101-114.

Examples

data(Battig)
## maybe str(Battig) ; plot(Battig) ...

# select items from several categories
cats <- c("fish", "bird", "flower", "tree")
for (c in cats) {
  cat("\nCategory:", c, "\n")  
  print(pickList(subset(Battig, catname==c), nitems=5))
}
#> 
#> Category: fish 
#>      list     word catnum catname syl letters freq frequency rank rfreq
#> 4635    1     shad     52    fish   1       4   26         0   28  3.65
#> 4623    1     carp     52    fish   1       4   61         0   16  3.14
#> 4609    1     bass     52    fish   1       4  195        16    2  3.81
#> 4631    1 bluefish     52    fish   2       8   30         1   24  5.11
#> 4638    1  haddock     52    fish   2       7   22         1   31  2.88
#> 
#> Category: bird 
#>      list      word catnum catname syl letters freq frequency rank rfreq
#> 2991    1   chicken     37    bird   2       7   40        37   21  6.00
#> 2998    1    falcon     37    bird   2       6   28         4   28  4.00
#> 2979    1 parrakeet     37    bird   3       9  115         1    9  4.61
#> 3001    1     finch     37    bird   1       5   20         0   31  5.10
#> 3000    1  pheasant     37    bird   2       8   22         1   30  5.10
#> 
#> Category: flower 
#>      list      word catnum catname syl letters freq frequency rank rfreq
#> 4239    1     pansy     48  flower   2       5  108         6    9  3.70
#> 4245    1     lilac     48  flower   2       5   45         4   15  5.22
#> 4257    1 buttercup     48  flower   3       9   17         0   27  5.46
#> 4262    1     aster     48  flower   2       5   12         1   32  6.60
#> 4258    1  marigold     48  flower   3       8   17         0   28  5.00
#> 
#> Category: tree 
#>      list           word catnum catname syl letters freq frequency rank rfreq
#> 4406    1           pine     50    tree   1       4  214        14    3  3.97
#> 4434    1     cottonwood     50    tree   3      10   14         0   31  3.50
#> 4438    1          holly     50    tree   2       5   11         0   35  4.36
#> 4413    1        redwood     50    tree   2       7   71         2   10  4.78
#> 4415    1 weeping willow     50    tree   4      13   67         0   12  4.79

# or, using sapply():
sapply(cats, function(c) pickList(subset(Battig, catname==c), nitems=5), simplify=FALSE)
#> $fish
#>      list     word catnum catname syl letters freq frequency rank rfreq
#> 4654    1  piranha     52    fish   3       7   10         0   47  5.33
#> 4616    1 goldfish     52    fish   2       8  127         1    9  3.77
#> 4619    1    whale     52    fish   1       5   78         0   12  4.58
#> 4636    1     spot     52    fish   1       4   25        57   29  5.04
#> 4623    1     carp     52    fish   1       4   61         0   16  3.14
#> 
#> $bird
#>      list      word catnum catname syl letters freq frequency rank rfreq
#> 3015    1 chickadee     37    bird   3       9   11         0   45  7.00
#> 2990    1   swallow     37    bird   2       7   41        10   20  4.64
#> 2978    1    canary     37    bird   3       6  134         0    8  4.03
#> 2983    1    oriole     37    bird   3       6   77        11   13  3.19
#> 3001    1     finch     37    bird   1       5   20         0   31  5.10
#> 
#> $flower
#>      list               word catnum catname syl letters freq frequency rank
#> 4242    1           daffodil     48  flower   3       8   87         1   12
#> 4264    1           camellia     48  flower   3       8   10         1   34
#> 4251    1 lily-of-the-valley     48  flower   6      15   29         0   21
#> 4239    1              pansy     48  flower   2       5  108         6    9
#> 4257    1          buttercup     48  flower   3       9   17         0   27
#>      rfreq
#> 4242  3.91
#> 4264  2.88
#> 4251  5.27
#> 4239  3.70
#> 4257  5.46
#> 
#> $tree
#>      list       word catnum catname syl letters freq frequency rank rfreq
#> 4425    1      cedar     50    tree   2       5   37         1   22  5.14
#> 4434    1 cottonwood     50    tree   3      10   14         0   31  3.50
#> 4429    1     orange     50    tree   2       6   29        23   26  7.72
#> 4411    1    dogwood     50    tree   2       7   84         1    8  4.89
#> 4420    1   sycamore     50    tree   3       8   46         0   17  4.39
#>