Battig - Montague Categorized Word Norms

This dataset comprises a ranked list of 5231 words listed in 56 taxonomic categories by people who were asked to list as many exemplars of a given category ("a precious stone", "a unit of time", "a fruit", "a color", etc.). Participants had 30s to generate as many responses to each category as possible, after which time the next category name was presented.

Included in this dataset are all words from the Battig and Montague (1969) norms listed with freq > 1.

Usage

data(Battig)

Format

A data frame with 5231 observations on the following 9 variables.

word

a character vector

catnum

category number, a factor

catname

category name, a factor

syl

number of syllables

letters

number of letters

freq

Frequency of response

frequency

Kucera-Francis word frequency

rank

rank of freq within the category

rfreq

rated frequency

Details

In our original dataset, words were truncated at 18 characters, so some are incomplete.

Source

Battig, W.F. & Montague, W.E. (1968). Category norms for verbal items in 56 categories: A replication and extension of the Connecticut norms using University of Maryland and Illinois students (Tech. Rep.) University of Colorado, Boulder, CO (1968)

Battig, W.F. & Montague, W.E. (1969). Category norms for verbal items in 56 categories: A replication and extension of the Connecticut norms. Journal of Experimental Psychology, 80 (1969), pp. 1-46

References

Joelson, J. M. & Hermann, D. J., Properties of categories in semantic | memory, Amer Journal of Psychology, 1978, 91, 101-114.

Examples

data(Battig)
## maybe str(Battig) ; plot(Battig) ...

# select items from several categories
cats <- c("fish", "bird", "flower", "tree")
for (c in cats) {
  cat("\nCategory:", c, "\n")  
  print(pickList(subset(Battig, catname==c), nitems=5))
}
#> 
#> Category: fish 
#>      list     word catnum catname syl letters freq frequency rank rfreq
#> 4645    1  sardine     52    fish   2       7   18         2   38  4.85
#> 4609    1     bass     52    fish   1       4  195        16    2  3.81
#> 4646    1 blowfish     52    fish   2       8   17         1   39  4.88
#> 4626    1    guppy     52    fish   2       5   53         0   19  4.38
#> 4653    1  dolphin     52    fish   2       7   11         1   46  6.70
#> 
#> Category: bird 
#>      list     word catnum catname syl letters freq frequency rank rfreq
#> 2992    1     dove     37    bird   1       4   36         4   22  4.32
#> 3012    1     swan     37    bird   1       4   14         3   42  5.63
#> 3005    1 flamingo     37    bird   3       8   16         0   35  5.00
#> 3000    1 pheasant     37    bird   2       8   22         1   30  5.10
#> 2994    1      owl     37    bird   1       3   36         2   24  5.57
#> 
#> Category: flower 
#>      list             word catnum catname syl letters freq frequency rank rfreq
#> 4252    1       snapdragon     48  flower   3      10   26         1   22  5.19
#> 4232    1            tulip     48  flower   2       5  209         4    2  3.46
#> 4254    1 black-eyed susan     48  flower   4      14   21         0   24  4.33
#> 4247    1            peony     48  flower   2       5   38         3   17  4.33
#> 4239    1            pansy     48  flower   2       5  108         6    9  3.70
#> 
#> Category: tree 
#>      list    word catnum catname syl letters freq frequency rank rfreq
#> 4422    1  poplar     50    tree   2       6   45         1   19  3.96
#> 4412    1  spruce     50    tree   1       6   74         5    9  4.61
#> 4419    1    pear     50    tree   1       4   49         6   16  5.47
#> 4411    1 dogwood     50    tree   2       7   84         1    8  4.89
#> 4407    1     elm     50    tree   1       3  210        13    4  2.64

# or, using sapply():
sapply(cats, function(c) pickList(subset(Battig, catname==c), nitems=5), simplify=FALSE)
#> $fish
#>      list     word catnum catname syl letters freq frequency rank rfreq
#> 4621    1      cod     52    fish   1       3   69         6   14  2.74
#> 4649    1  walleye     52    fish   2       7   16         0   42  5.00
#> 4611    1  herring     52    fish   2       7  161         2    4  2.99
#> 4628    1   marlin     52    fish   2       6   33         2   21  5.40
#> 4631    1 bluefish     52    fish   2       8   30         1   24  5.11
#> 
#> $bird
#>      list        word catnum catname syl letters freq frequency rank rfreq
#> 2979    1   parrakeet     37    bird   3       9  115         1    9  4.61
#> 2972    1     sparrow     37    bird   2       7  237         1    2  4.53
#> 2999    1         jay     37    bird   1       3   24        15   29  4.00
#> 2986    1 hummingbird     37    bird   3      11   50         0   16  4.69
#> 2977    1    bluebird     37    bird   2       8  138         1    7  3.36
#> 
#> $flower
#>      list     word catnum catname syl letters freq frequency rank rfreq
#> 4238    1     lily     48  flower   2       4  108         1    8  4.17
#> 4261    1 magnolia     48  flower   4       8   13         1   31  7.00
#> 4264    1 camellia     48  flower   3       8   10         1   34  2.88
#> 4258    1 marigold     48  flower   3       8   17         0   28  5.00
#> 4240    1  petunia     48  flower   3       7  101         0   10  2.93
#> 
#> $tree
#>      list       word catnum catname syl letters freq frequency rank rfreq
#> 4419    1       pear     50    tree   1       4   49         6   16  5.47
#> 4417    1        fir     50    tree   1       3   60         2   14  5.38
#> 4434    1 cottonwood     50    tree   3      10   14         0   31  3.50
#> 4407    1        elm     50    tree   1       3  210        13    4  2.64
#> 4428    1      beech     50    tree   1       5   29         6   25  5.05
#>