Select Items from a Word Pool in Given Ranges

This is a convenience function to provide the capability to select items from a given word pool, with restrictions on the range of any numeric variables.

Usage

pickList(data, ranges, nitems = 10, nlists = 1, replace = FALSE)

Arguments

data

A data.frame, typically a word list like Paivio or TWP

ranges

A data.frame of two rows, and with column names corresponding to a subset of the column names in data. The two rows give the minimum and maximum values, respectively, of variables in data. Alternatively, you can supply a named list containing the minimum and maximum values for one or more variables in data.

nitems

Number of items per list

nlists

Number of lists

replace

A logical value, indicating whether the sampling of items (rows) of data is to allow sampling with replacement.

Value

A data frame of the same shape as data containing the selected items prefixed by the list number.

Details

sample will generate an error if fewer than nitems * nlists items are within the specified ranges and replace=FALSE.

References

A related word list generator: Friendly, M. Word list generator. http://datavis.ca/online/paivio/

See also

Author

Michael Friendly

Examples

data(Paivio)
# 2 lists, no selection on any variables
pickList(Paivio, nlists=2)
#>     list itmno          word imagery concreteness meaningfulness frequency syl
#> 269    1   269       dreamer    5.33         5.10           5.04         4   2
#> 386    1   386          golf    6.70         6.10           8.16        26   1
#> 411    1   411  headquarters    5.03         5.94           5.08        23   3
#> 323    1   323         fault    2.83         2.87           4.80        50   1
#> 194    1   194      contract    4.50         5.24           6.32        50   2
#> 832    1   832         table    6.50         7.00           7.60       100   1
#> 337    1   337         flood    6.33         6.62           6.36        50   1
#> 369    1   369         geese    6.53         6.83           7.48        26   1
#> 313    1   313 extermination    4.03         3.13           4.96         2   5
#> 135    1   135           car    6.87         7.00           6.38       100   1
#> 566    2   566        memory    3.10         1.78           5.00        50   3
#> 445    2   445         idiom    3.00         3.44           4.60         1   3
#> 193    2   193       context    2.13         2.73           4.44         1   2
#> 277    2   277  eccentricity    3.77         2.27           4.04         2   5
#> 684    2   684       prairie    6.47         6.80           8.16        25   2
#> 524    2   524    letterhead    5.13         6.45           4.92         1   3
#> 232    2   232      delirium    4.80         2.79           5.00         2   4
#> 590    2   590        morgue    6.03         6.48           6.56         2   2
#> 671    2   671        pledge    3.63         2.93           5.92        27   1
#> 140    2   140   caterpillar    6.57         6.96           6.88        19   4
#>     letters freerecall
#> 269       7      0.500
#> 386       4      0.469
#> 411      12      0.469
#> 323       5      0.188
#> 194       8      0.313
#> 832       5      0.313
#> 337       5      0.563
#> 369       5      0.406
#> 313      13      0.594
#> 135       3      0.656
#> 566       6      0.656
#> 445       5      0.281
#> 193       7      0.313
#> 277      12      0.594
#> 684       7      0.438
#> 524      10      0.469
#> 232       8      0.313
#> 590       6      0.500
#> 671       6      0.219
#> 140      11      0.469

# Define ranges for low and high on imagery, concreteness, meaningfulness
# These go from low - median, and median-high on each variable
vars <- 3:5
(low <- as.data.frame(apply(Paivio[,vars], 2, fivenum))[c(1,3),])
#>   imagery concreteness meaningfulness
#> 1    1.63         1.18           1.92
#> 3    5.17         5.72           5.92
(high <- as.data.frame(apply(Paivio[,vars], 2, fivenum))[c(3,5),])
#>   imagery concreteness meaningfulness
#> 3    5.17         5.72           5.92
#> 5    6.90         7.70           9.22

# select two lists of 10 low/high imagery items
lowI <- pickList(Paivio, low[,"imagery", drop=FALSE], nitems=10, nl=2)
highI <- pickList(Paivio, high[,"imagery", drop=FALSE], nitems=10, nl=2)

# compare means
colMeans(lowI[,c(4:8)])
#>        imagery   concreteness meaningfulness      frequency            syl 
#>         3.6345         3.4555         5.3855        30.9000         2.4500 
colMeans(highI[,c(4:8)])
#>        imagery   concreteness meaningfulness      frequency            syl 
#>         6.0650         6.0000         6.5165        34.8000         2.0000 

# using a list of ranges
L <- list(imagery=c(1,5), concreteness=c(1,4))
pickList(Paivio, L)
#>     list itmno        word imagery concreteness meaningfulness frequency syl
#> 300    1   300     essence    2.33         1.66           3.28         8   2
#> 444    1   444        idea    2.20         1.42           4.88       100   2
#> 6      1     6  aberration    2.27         2.98           2.64         0   4
#> 292    1   292  engagement    4.70         3.46           5.76        24   3
#> 168    1   168    clemency    2.90         1.97           3.94         1   3
#> 191    1   191    conquest    4.03         3.11           5.81        26   2
#> 901    1   901      vision    4.73         4.00           6.39        45   2
#> 735    1   735      ritual    4.70         3.70           5.12         4   3
#> 283    1   283 elaboration    2.63         2.18           4.36         1   5
#> 454    1   454    incident    2.90         3.00           4.16        29   3
#>     letters freerecall
#> 300       7         NA
#> 444       4      0.313
#> 6        10      0.375
#> 292      10      0.500
#> 168       8      0.344
#> 191       8      0.094
#> 901       6      0.344
#> 735       6      0.281
#> 283      11      0.219
#> 454       8      0.156