Select Items from a Word Pool in Given Ranges

This is a convenience function to provide the capability to select items from a given word pool, with restrictions on the range of any numeric variables.

Usage

pickList(data, ranges, nitems = 10, nlists = 1, replace = FALSE)

Arguments

data

A data.frame, typically a word list like Paivio or TWP

ranges

A data.frame of two rows, and with column names corresponding to a subset of the column names in data. The two rows give the minimum and maximum values, respectively, of variables in data. Alternatively, you can supply a named list containing the minimum and maximum values for one or more variables in data.

nitems

Number of items per list

nlists

Number of lists

replace

A logical value, indicating whether the sampling of items (rows) of data is to allow sampling with replacement.

Value

A data frame of the same shape as data containing the selected items prefixed by the list number.

Details

sample will generate an error if fewer than nitems * nlists items are within the specified ranges and replace=FALSE.

References

A related word list generator: Friendly, M. Word list generator. http://datavis.ca/online/paivio/

See also

Author

Michael Friendly

Examples

data(Paivio)
# 2 lists, no selection on any variables
pickList(Paivio, nlists=2)
#>     list itmno        word imagery concreteness meaningfulness frequency syl
#> 106    1   106     boulder    6.13         6.96           5.88        50   2
#> 9      1     9       abyss    5.17         5.40           5.88         5   2
#> 426    1   426   honeycomb    6.23         6.50           5.78         4   3
#> 322    1   322     fatigue    5.07         4.28           3.88        19   2
#> 810    1   810       stone    6.23         6.96           7.24       100   1
#> 617    1   617    occasion    2.53         3.22           5.00        50   3
#> 131    1   131       candy    6.63         6.56           6.39        32   2
#> 719    1   719 recognition    3.60         2.00           5.04        21   4
#> 862    1   862     tribute    3.57         3.12           5.52        22   2
#> 626    1   626   orchestra    6.77         6.55           6.76        10   3
#> 386    2   386        golf    6.70         6.10           8.16        26   1
#> 714    2   714    railroad    6.27         6.76           6.60       100   2
#> 820    2   820  substitute    3.13         4.56           4.48        29   3
#> 158    2   158    chloride    4.20         6.28           5.75         2   2
#> 346    2   346        fork    6.57         6.94           7.08        31   1
#> 536    2   536        lord    4.63         4.18           6.68       100   1
#> 610    2   610         nun    6.67         6.76           6.60         9   1
#> 420    2   420    hillside    6.30         6.59           6.92        20   2
#> 486    2   486    jealousy    4.72         1.77           4.88        17   3
#> 301    2   301  evangelist    4.97         5.95           5.28         2   4
#>     letters freerecall
#> 106       7      0.500
#> 9         5      0.438
#> 426       9      0.281
#> 322       7      0.438
#> 810       5      0.531
#> 617       8      0.219
#> 131       5      0.500
#> 719      11      0.500
#> 862       7      0.281
#> 626       9      0.344
#> 386       4      0.469
#> 714       8      0.313
#> 820      10      0.125
#> 158       8      0.281
#> 346       4      0.469
#> 536       4      0.750
#> 610       3      0.563
#> 420       8      0.656
#> 486       8      0.438
#> 301      10      0.625

# Define ranges for low and high on imagery, concreteness, meaningfulness
# These go from low - median, and median-high on each variable
vars <- 3:5
(low <- as.data.frame(apply(Paivio[,vars], 2, fivenum))[c(1,3),])
#>   imagery concreteness meaningfulness
#> 1    1.63         1.18           1.92
#> 3    5.17         5.72           5.92
(high <- as.data.frame(apply(Paivio[,vars], 2, fivenum))[c(3,5),])
#>   imagery concreteness meaningfulness
#> 3    5.17         5.72           5.92
#> 5    6.90         7.70           9.22

# select two lists of 10 low/high imagery items
lowI <- pickList(Paivio, low[,"imagery", drop=FALSE], nitems=10, nl=2)
highI <- pickList(Paivio, high[,"imagery", drop=FALSE], nitems=10, nl=2)

# compare means
colMeans(lowI[,c(4:8)])
#>        imagery   concreteness meaningfulness      frequency            syl 
#>         3.7575         3.4060         5.3955        15.9000         2.6500 
colMeans(highI[,c(4:8)])
#>        imagery   concreteness meaningfulness      frequency            syl 
#>         5.9585         6.3710         6.5700        38.1000         1.9000 

# using a list of ranges
L <- list(imagery=c(1,5), concreteness=c(1,4))
pickList(Paivio, L)
#>     list itmno         word imagery concreteness meaningfulness frequency syl
#> 880    1   880       upkeep    3.07         2.50           4.76         2   2
#> 710    1   710     quantity    3.47         3.32           4.17        50   3
#> 865    1   865      trouble    3.53         2.25           5.08       100   2
#> 413    1   413      hearing    3.77         3.57           6.64        13   2
#> 763    1   763       series    4.47         3.88           5.36        50   2
#> 885    1   885       vanity    3.83         1.77           4.92        20   3
#> 689    1   689     prestige    3.67         1.73           4.56         7   2
#> 575    1   575      miracle    3.33         2.25           5.60        24   3
#> 414    1   414       heaven    4.13         2.75           6.52       100   2
#> 195    1   195 contribution    3.67         3.33           5.44        19   4
#>     letters freerecall
#> 880       6      0.281
#> 710       8      0.313
#> 865       7      0.250
#> 413       7      0.156
#> 763       6      0.281
#> 885       6      0.344
#> 689       8      0.188
#> 575       7      0.188
#> 414       6      0.656
#> 195      12      0.125